backend/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.21.22
botocore==1.24.22
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Collectfast==2.2.0
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.6
django-allauth==0.50.0
django-cors-headers==3.11.0
django-health-check==3.16.5
django-ipware==4.0.2
django-redis==5.2.0
django-sendgrid-v5==1.2.0
django-silk==4.3.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
geoip2==2.9.0
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
jmespath==1.0.0
jsonschema==4.4.0
kombu==5.2.4
MarkupSafe==2.1.1
maxminddb==2.2.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.3
redis==4.2.2
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
sendgrid==6.9.7
sentry-sdk==1.5.7
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tldextract==3.2.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
